Reduce webapp page load time
