+++ webapp/src/test/app/pages.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { describe, it, expect, vi } from "vitest";
 import HomePage from "@/app/page";
-import SongsetsPage from "@/app/songsets/page";
+import { SongsetsClient } from "@/app/songsets/SongsetsClient";
 import SettingsPage from "@/app/settings/page";
 
 vi.mock("next/navigation", () => ({
@@ -24,7 +24,7 @@ describe("HomePage", () => {
 
 describe("SongsetsPage", () => {
   it("renders heading", () => {
-    render(<SongsetsPage />);
+    render(<SongsetsClient initialData={{ songsets: [], total: 0 }} />);
     expect(screen.getByRole("heading", { name: /songsets/i })).toBeInTheDocument();
   });
 });
