+++ webapp/src/test/app/render-page.test.tsx
@@ -1,82 +1,46 @@
-import { describe, it, expect, vi, beforeEach } from "vitest"
-import { render, screen, waitFor } from "@testing-library/react"
-import RenderPage from "@/app/songsets/[id]/render/page"
+import { describe, it, expect, vi } from "vitest"
+import { render, screen } from "@testing-library/react"
+import { RenderPageClient } from "@/app/songsets/[id]/render/RenderPageClient"
 
 const mockPush = vi.fn()
 vi.mock("next/navigation", () => ({
-  useParams: () => ({ id: "test-songset" }),
   useRouter: () => ({
     push: mockPush,
   }),
 }))
 
-const mockSongsetData = {
-  id: "test-songset",
-  name: "Sunday Worship",
-  description: "Easter service",
-  items: [{ markedLineCount: 2 }, { markedLineCount: 1 }],
-  latestRenderJobId: null,
-  lastFailedRenderJobId: null,
-  lastCompletedRenderJobId: null,
-  renderState: "unrendered",
-}
+describe("RenderPageClient", () => {
+  it("renders server-loaded songset data", () => {
+    render(
+      <RenderPageClient
+        songsetId="test-songset"
+        initialSongset={{
+          id: "test-songset",
+          name: "Sunday Worship",
+          description: "Easter service",
+          markedLineCount: 0,
+          renderState: "unrendered",
+          songTitles: [],
+          lastCompletedRenderJobId: null,
+        }}
+        initialLatestJob={null}
+        initialPreviousCompletedJob={null}
+        initialRenderData={{
+          audioEnabled: true,
+          videoEnabled: true,
+          template: "dark",
+          resolution: "720p",
+          fontSizePreset: "M",
+          fontFamily: "noto_serif_tc",
+          includeTitleCard: false,
+          titleCardDurationSeconds: 10,
+          titleCardLines: [],
+          offlineEnabled: false,
+        }}
+      />
+    )
 
-const mockSettingsData = {
-  settings: {
-    userId: 1,
-    offlineAutoCache: true,
-    defaultGapBeats: 2.0,
-    defaultVideoTemplate: "dark",
-    defaultResolution: "720p",
-    lyricsLoopWindowSeconds: 3.0,
-    defaultFontSizePreset: "M",
-    defaultFontFamily: "noto_serif_tc",
-    defaultKeyShiftSemitones: 0,
-    timingReviewFont: "sans",
-  },
-}
-
-describe("RenderPage", () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockPush.mockClear()
+    expect(screen.getByRole("heading", { name: /render/i })).toBeInTheDocument()
+    expect(screen.getByText("Sunday Worship")).toBeInTheDocument()
   })
-
-  it("renders loading state initially", () => {
-    global.fetch = vi.fn().mockImplementation((url: string) => {
-      if (url.includes("/api/songsets/")) {
-        return Promise.resolve({ ok: true, status: 200, json: () => Promise.resolve(mockSongsetData) })
-      }
-      if (url.includes("/api/settings")) {
-        return Promise.resolve({ ok: true, status: 200, json: () => Promise.resolve(mockSettingsData) })
-      }
-      return Promise.resolve({ ok: false, status: 404 })
-    })
-    render(<RenderPage />)
-    expect(screen.getByRole("status")).toBeInTheDocument()
-  })
-
-  it("redirects to login when unauthorized", async () => {
-    global.fetch = vi.fn().mockResolvedValue({
-      ok: false,
-      status: 401,
-    })
-    render(<RenderPage />)
-    await waitFor(() => {
-      expect(mockPush).toHaveBeenCalledWith("/login")
-    }, { timeout: 10000 })
-  }, 15000)
-
-  it("shows error when songset not found", async () => {
-    global.fetch = vi.fn().mockImplementation((url: string) => {
-      if (url.includes("/api/songsets/")) {
-        return Promise.resolve({ ok: false, status: 404 })
-      }
-      return Promise.resolve({ ok: true, status: 200, json: () => Promise.resolve(mockSettingsData) })
-    })
-    render(<RenderPage />)
-    await waitFor(() => {
-      expect(screen.getByText(/songset not found/i)).toBeInTheDocument()
-    }, { timeout: 10000 })
-  }, 15000)
 })
